pyproject.toml
[project]
name = "HyperBoxLauncher"
version = "0.1.0"
description = "Launcher for my minecraft project"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
  "aiofiles>=23.2.1",
  "python-dotenv>=1.0.0",
  "loguru>=0.7.3",
  "minecraft-launcher-lib>=8.0",
  "pillow>=10.0.0",
  "psutil>=5.9.0",
  "pywebview==6.1",
  "tenacity>=9.1.2",
  "httpx>=0.28.1",
  "nuitka>=2.8.8",
]